pimd: Smarten up pim_nexthop_lookup
There exists cases where we have not properly configured
a RP yet, but we are getting callbacks for nocache.
This is generating allot of spam messages in the log.

In the case where we attempt to lookup INADDR_NONE
just say it fails without need to talk to anyone
since we'll never get a real answer from zebra
anyways.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>